Refactor to match naming scheme
