+++ MikuMikuLibrary/IO/Sections/Chritm/ChritmSetSection.cs
@@ -5,13 +5,13 @@ using MikuMikuLibrary.IO.Sections.IO;
 namespace MikuMikuLibrary.IO.Sections.Chritm
 {
     [Section("ITEM")]
-    public class AetSetSection : BinaryFileSection<ChritmSet>
+    public class AetSetSection : BinaryFileSection<CharacterItemTable>
     {
         public override SectionFlags Flags => SectionFlags.None;
 
         public override Encoding Encoding { get; } = Encoding.GetEncoding("utf-8");
 
-        public AetSetSection(SectionMode mode, ChritmSet data = null) : base(mode, data)
+        public AetSetSection(SectionMode mode, CharacterItemTable data = null) : base(mode, data)
         {
         }
     }
